LPS-99391 move logic into AccountDisplay
